Implement walkJsonDirSync
